requirements.txt
torch==2.0.1
torchvision==0.15.2
transformers>=4.33.0
numpy==1.24.3
Pillow>=9.5.0
huggingface-hub>=0.16.4
timm>=0.9.0
safetensors>=0.3.1
einops>=0.6.1
opencv-python-headless>=4.8.0
scipy>=1.11.2
ftfy>=6.1.1
regex>=2023.8.8
tqdm>=4.66.1
